+++ internal/mithril-persistence/src/database/provider/block_range_root/insert_block_range.rs
@@ -3,11 +3,9 @@ use std::iter::repeat;
 use sqlite::Value;
 
 use mithril_common::StdResult;
-use mithril_persistence::sqlite::{
-    Provider, SourceAlias, SqLiteEntity, SqliteConnection, WhereCondition,
-};
 
 use crate::database::record::BlockRangeRootRecord;
+use crate::sqlite::{Provider, SourceAlias, SqLiteEntity, SqliteConnection, WhereCondition};
 
 /// Query to insert [BlockRangeRootRecord] in the sqlite database
 pub struct InsertBlockRangeRootProvider<'client> {