+++ internal/mithril-persistence/src/database/provider/cardano_transaction/mod.rs
@@ -1,5 +1,7 @@
+mod delete_cardano_transaction;
 mod get_cardano_transaction;
 mod insert_cardano_transaction;
 
+pub use delete_cardano_transaction::*;
 pub use get_cardano_transaction::*;
 pub use insert_cardano_transaction::*;