+++ internal/mithril-persistence/src/database/record/mod.rs
@@ -1,0 +1,9 @@
+//! Shared database records
+
+mod block_range_root;
+mod cardano_transaction;
+mod interval_without_block_range_root;
+
+pub use block_range_root::*;
+pub use cardano_transaction::*;
+pub use interval_without_block_range_root::*;