+++ mithril-aggregator/src/services/cardano_transactions_importer.rs
@@ -19,12 +19,6 @@ pub trait TransactionStore: Send + Sync {
     /// Get the highest known transaction beacon
     async fn get_highest_beacon(&self) -> StdResult<Option<ImmutableFileNumber>>;
 
-    /// Get stored transactions up to the given beacon
-    async fn get_up_to(
-        &self,
-        immutable_file_number: ImmutableFileNumber,
-    ) -> StdResult<Vec<CardanoTransaction>>;
-
     /// Store list of transactions
     async fn store_transactions(&self, transactions: Vec<CardanoTransaction>) -> StdResult<()>;
 
@@ -189,8 +183,8 @@ mod tests {
     };
     use mithril_common::crypto_helper::MKTree;
     use mithril_common::entities::{BlockNumber, BlockRangesSequence};
+    use mithril_persistence::database::repository::CardanoTransactionRepository;
 
-    use crate::database::repository::CardanoTransactionRepository;
     use crate::database::test_helper::cardano_tx_db_connection;
 
     use super::*;